styling + hover effect + gallery + description
I changed the styling, added a gradient from top to bottom and a text colour. Hover effect was added to a link. I replaced random gallery photos with the winners' photos of Comedy Wildlife Photography Awards from various years, I added the description to each image and a link to the official website.